test: split test-crypto-dh to avoid timeout on slow machines in the CI
Locally this speeds up running test-crypto-dh* from 7s to 2s. This was
previously timing out in CI (took more than 2 minutes) so should see
a bigger gap in the CI.

PR-URL: https://github.com/nodejs/node/pull/49492
Refs: https://github.com/nodejs/node/issues/49202
Refs: https://github.com/nodejs/reliability/issues/655
